+++ CHANGELOG.md
@@ -28,7 +28,7 @@
 - Adding or removing a career will adjust the amount of renown and wealth your character has.
 - Added a Renown and Wealth field to characters. (#279)
 - Added age, height, and weight info to the Biography tab for characters.
-- Expanded end of encounter prompt to reset heroic resources & temporary stamina (#612)
+- Expanded end of encounter prompt to reset heroic resources & temporary stamina. (#612)
 - Added name and image inputs to pseudodocument creation, e.g. Power Roll Effects. (#718)
 
 ### Changed